+++ neuroimg/pipeline/config/localconfig.yaml
@@ -17,7 +17,7 @@ derivatives_output_dir:
 patients:
   [
 #    "PublicDataset",
-    'la02',
+    'la03',
 #    "la04",
 #      'efri02',
 #     'efri06',
